Decouple exec_meta_search from social_links
