+++ witch_craft/src/modules/osint/lookup.rs
@@ -1,5 +1,3 @@
-use reqwest::blocking::Client;
-
 use crate::core::core::*;
 
 pub fn search_ans(argsv: &[String]) -> i32 {
@@ -95,114 +93,5 @@ pub fn cinsscore(argsv: &[String]) -> i32 {
     }
 
     raise("Nothing found, must be legit UwU", "message");
-    return 0;
-}
-
-use headless_chrome::{Browser, LaunchOptionsBuilder};
-
-pub fn social_links(argsv: &[String]) -> i32 {
-    let keyword = search_value("keyword", argsv);
-    // name - url - filter
-    let social_links = [
-        ("youtube", "https://www.youtube.com/@@@keyword", ""),
-        ("gitlab", "https://gitlab.com/@@keyword", ""),
-        ("github", "https://github.com/@@keyword", ""),
-        ("bitbucket", "https://bitbucket.org/@@keyword", ""),
-        (
-            "facebook",
-            "https://facebook.com/@@keyword",
-            "This content isn't available right now",
-        ),
-        ("slideshare", "https://slideshare.net/@@keyword", ""),
-        (
-            "linkedin.corp",
-            "https://linkedin.com/company/@@keyword",
-            "",
-        ),
-        ("linkedin.user", "https://linkedin.com/in/@@keyword", ""),
-        ("myspace", "https://myspace.com/@@keyword", ""),
-        (
-            "instagram",
-            "https://instagram.com/@@keyword",
-            "Sorry, this page isn't available.",
-        ),
-        (
-            "medium",
-            "https://medium.com/@@@keyword",
-            "Out of nothing, something.",
-        ),
-        (
-            "twitch",
-            "https://twitch.tv/@@keyword",
-            "Sorry. Unless you've got a time machine, that content is unavailable.",
-        ),
-        ("mastodon", "https://mastodon.social/@@@keyword", ""),
-        (
-            "bsky",
-            "https://bsky.app/profile/@@keyword",
-            "Error: handle must be a valid handle",
-        ),
-        ("reddit", "https://www.reddit.com/user/@@keyword", ""),
-        (
-            "twitter",
-            "https://www.x.com/@@keyword",
-            "This account doesn’t exist",
-        ),
-        (
-            "xvideos",
-            "https://www.xvideos.com/profiles/@@keyword",
-            "THIS PROFILE DOESN'T EXIST !",
-        ),
-        (
-            "snapchat",
-            "https://www.snapchat.com/add/@@keyword",
-            "This content was not found",
-        ),
-    ];
-
-    for item in social_links {
-        let client = Client::new();
-        let browser = Browser::new(
-            LaunchOptionsBuilder::default()
-                .headless(true) // Run the browser in headless mode
-                .build()
-                .unwrap(),
-        )
-        .unwrap();
-        let tab = browser.new_tab().unwrap();
-        let url = item.1.replace("@@keyword", &keyword);
-
-        // Tab settings
-        tab.set_default_timeout(std::time::Duration::from_secs(30));
-        tab.navigate_to(&url).unwrap();
-        tab.wait_until_navigated().unwrap();
-
-        let content = tab.get_content().unwrap();
-
-        match client.get(&url).send() {
-            Ok(res) => {
-                if res.status().as_u16() == 200 {
-                    if item.2.is_empty() {
-                        raise(
-                            &format!("Found! {} {} at {}", &keyword, item.0, &url),
-                            "done",
-                        );
-                    } else {
-                        if content.contains(item.2) == false {
-                            raise(
-                                &format!("Found! {} {} at {}", &keyword, item.0, &url),
-                                "done",
-                            );
-                        }
-                    }
-                }
-            }
-            Err(err) => {
-                eprintln!("{:?}", err);
-                return 1;
-            }
-        }
-    }
-
     return 0;
 }
