+++ witch_craft/src/modules/osint/mod.rs
@@ -1,2 +1,4 @@
+pub mod data;
 pub mod lookup;
+pub mod meta_search;
 pub mod osint;
