+++ witch_craft/src/modules/osint/osint.rs
@@ -1,5 +1,6 @@
 use crate::core::types::Closure;
 use crate::modules::osint::lookup::*;
+use crate::modules::osint::meta_search::social_links;
 
 pub fn api() -> Closure {
     vec![
@@ -7,6 +8,6 @@ pub fn api() -> Closure {
         ("search.geoloc", search_geoloc),
         ("search.proxy", search_proxy),
         ("search.ipscore", cinsscore),
-        ("search.social", social_links),
+        ("search.meta", social_links),
     ]
 }
